Add a GitHub action to push NuGet packages when a release is created
This commit won't actually push anything - but it will allow
us to test things before we *do* manually create a release.

Once testing is completed, we'll want to use a (suitably encrypted)
NuGet API key to push to NuGet.

Note that this initially assumes we want to release all packages at
the same time. We'll want to change it when we reach GA for 2.0.0.

Signed-off-by: Jon Skeet <jonskeet@google.com>