Expose a CloudEventAttributeTypeOrdinal enum for CloudEventAttributeType
This is in the Core namespace as most users won't need it, and
instead of being exposed publicly on CloudEventAttributeType as a
property, it's available via a static method in a utility type in
the Core namespace. This effectively obscures it a little, while
still keeping it available to the formatter/protocol code it's
aimed at.

Signed-off-by: Jon Skeet <jonskeet@google.com>