fix: Change AMQP data handling
Rather than the AMQP code "knowing" how to handle different data
types, delegate that to an event formatter.

There's *still* a little bit of an ugly
not-quite-single-responsibility in terms of event formatters;
CloudEvent formats tend to be specified in terms of structured mode
messages, not binary mode. It's possible that we should really
separate out "data formatting" from everything else... but it does
get intertwined in structured mode formats.

This change fixes #104, but we certainly want more testing and
documentation for this code in general.

Signed-off-by: Jon Skeet <jonskeet@google.com>