fix: Validate that source is non-empty
URI references can be empty - we have a test proving that. But the source attribute is required to be non-empty, so we should validate that.

This bug fix is a breaking change for anyone who was previously creating (and then validating) CloudEvent instances with a present-but-empty source.
However, their events were already broken, and we'd expect them to cause issues with other consumers (at least any performing comprehensive validation).

Signed-off-by: Jon Skeet <jonskeet@google.com>