chore: Add editorconfig
(This will be applied in subsequent commits.)

One potentially-controversial aspect of this editorconfig is the use of a space after casting. The code is currently very inconsistent, but has a space after the cast more often than not (the diff is smaller with it set to true than false)... it's also consistent with the style I personally use in multiple other repositories. While I'm the main contributor, it probably makes sense to stick to that.

Signed-off-by: Jon Skeet <jonskeet@google.com>